+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.0.6
+
+- Added `time_expr_parser` to `Time_expr` API
+
+- Added `duration_expr_parser` to `Duration` API
+
+- Removed branching time point expression
+
+- Adjusted grammar for branching time slot expression
+
+- Replaced `skip_space` with `skip_white` so empty spaces across lines are skipped as well
+
+- Added syntax (`tzoffset=...`) for searching in a specific timezone offset
+
 ## 0.0.5
 
 - Fixed leap year definition